+++ R/roxygen_description.R
@@ -1,16 +1,18 @@
 # This add the lines
 # importFrom(...,...)
+# useDynLib(IBMPopSim, .registration=TRUE)
+# exportPattern("^[[:alpha:]]+")
 # to the NAMESPACE file
 # when calling roxygen2::roxygenize()
 
-#' @importFrom Rcpp sourceCpp
-#' @importFrom reshape melt cast
+#' @useDynLib IBMPopSim, .registration=TRUE
+
+#' @exportPattern "^[[:alpha:]]+"
+
+#' @importFrom Rcpp sourceCpp evalCpp
 #' @importFrom rlang .data
-#' @importFrom dplyr filter count transmute mutate group_by group_by_at select select_at summarise n %>% everything full_join
-#' @importFrom purrr reduce
-#' @importFrom tidyr complete
+#' @importFrom dplyr filter count mutate group_by group_by_at select select_at summarise n %>% everything
 #' @importFrom stats runif
-#' @importFrom utils read.table
 #' @importFrom ggplot2 ggplot aes geom_line geom_point geom_step geom_bar ggtitle theme xlab ylab coord_flip scale_x_discrete scale_y_continuous scale_fill_manual geom_col element_text
 #' @import readr
 #' @import checkmate