+++ inst/include/event_individual_template.cpp
@@ -1,7 +1,7 @@
 // Event "_NAME_"
 struct event_J_ : public event {
-    double __lambda(unsigned k, population const & pop, double t, context & cntxt) const {
-        individual const & I = pop[k];
+    double __lambda(unsigned _k, population const & pop, double t, context & cntxt) const {
+        individual const & I = pop[_k];
         double result = 0;
         _INTENSITY_CODE_
         return result;
@@ -10,20 +10,20 @@ struct event_J_ : public event {
         double result = cntxt.__intensity_bound[_J_] * pop.nAlive;
         return result;
     }
-    void __kernel(unsigned k, population & pop, double t, context & cntxt) const {
-        individual & I = pop[k];
+    void __kernel(unsigned _k, population & pop, double t, context & cntxt) const {
+        individual & I = pop[_k];
         _KERNEL_METHOD_
     }
     bool __apply(double t, population & pop, context & cntxt) const {
-        unsigned k = pop.pick_potentially_alive(t, cntxt);
-        if (pop.is_alive(k, t)) {
+        unsigned _k = pop.pick_potentially_alive(t, cntxt);
+        if (pop.is_alive(_k, t)) {
             // Make a reject strategy
             double theta = CUnif() * __lambda_bound(pop, cntxt) / pop.nAlive;
-            if (theta < __lambda(k, pop, t, cntxt)) {
-                __kernel(k, pop, t, cntxt);
+            if (theta < __lambda(_k, pop, t, cntxt)) {
+                __kernel(_k, pop, t, cntxt);
                 return true;
             }
-        } else pop.kill(k, t);
+        } else pop.kill(_k, t);
         return false;
     }
 };