+++ inst/include/event_poisson_template.cpp
@@ -1,6 +1,6 @@
 // Event "_NAME_"
 struct event_J_ : public event {
-    double __lambda(unsigned k, population const & pop, double t, context & cntxt) const {
+    double __lambda(unsigned _k, population const & pop, double t, context & cntxt) const {
         double result = cntxt.__intensity_bound[_J_];
         return result;
     }
@@ -8,13 +8,13 @@ struct event_J_ : public event {
         double result = cntxt.__intensity_bound[_J_] / cntxt.__num_threads;
         return result;
     }
-    void __kernel(unsigned k, population & pop, double t, context & cntxt) const {
-        individual & I = pop[k];
+    void __kernel(unsigned _k, population & pop, double t, context & cntxt) const {
+        individual & I = pop[_k];
         _KERNEL_METHOD_
     }
     bool __apply(double t, population & pop, context & cntxt) const {
-        unsigned k = _PICK_IFNOT_ENTRY_;
-        __kernel(k, pop, t, cntxt);
+        unsigned _k = _PICK_IFNOT_ENTRY_;
+        __kernel(_k, pop, t, cntxt);
         return true;
     }
 };