+++ NEWS.md
@@ -1,6 +1,15 @@
-# IBMPopSim 0.3.1
+# IBMPopSim 0.4.3
 
 ## Bug fixes
 - Add figures in interaction vignette
-- Fix precompile.R for png figures
 - Fix run test in mk_model and plot_pyramid functions
+- Add reserved names for characteristics and parameters : I, J, t, pop, newI, id, entry, out
+- Hide plot commands from vignettes
+- Add default value for "out" column for entry events 
+- Add default value for "entry" column for birth event
+- Add characteristics tracking to merge_pop_withid
+
+## New features
+- Convert exposure_table and death_table to Rcpp for more speed
+- Reduce dependencies
+